# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "description": "An API for managing users in a database",
    "version": "1.0.0"
  },
  "paths": {
    "/conversations": {
      "post": {
        "tags": ["conversations"],
        "summary": "Create a new conversation",
        "description": "Create a conv after 2 user matched",
        "parameters": [
          {
            "name": "sender_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "receiver_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          }],
        "responses": {
          "200": {
            "description": "Successfully createad a conversation",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Conversations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve conversation from the database"
          }
        }
      }
    },
    "/conversations/{id}": {
      "delete": {
        "tags": ["conversations"],
        "summary": "Delete a conversation",
        "description": "Delete a conversation and all the messages linked to it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }],
        "responses": {
          "200": {
            "description": "Successfully deleted the conversation",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Conversations"
              }
            }
          },
          "500": {
            "description": "Failed to delete conversation from the database"
          }
        }
      }
    },
    "/conversations/{userId}": {
      "get": {
        "tags": ["conversations"],
        "summary": "Get all conversation of a user",
        "description": "Get all the conversations",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          }],
        "responses": {
          "200": {
            "description": "Successfully gotten the conversations",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Conversations"
              }
            }
          },
          "500": {
            "description": "Failed to get conversations from the database"
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get all notifications",
        "description": "Get all notifications from the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved all notifications",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notifications"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve notifications from the database"
          }
        }
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Insert a new notification",
        "description": "Create notification in the database",
        "parameters": [
          {
            "name": "sender_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "receiver_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "type",
            "in": "formData",
            "required": true,
            "type": "string"
          }],
        "responses": {
          "200": {
            "description": "Successfully created the notification",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Blocks"
              }
            }
          },
          "500": {
            "description": "Failed to create the notification in the database"
          }
        }
      }
    },
    "/notifications/{id}": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "get notifications of user",
        "description": "get notifications of user",
        "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer"
        }],
        "responses": {
          "200": {
            "description": "Successfully gotten all the notif of user",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notifications"
              }
            }
          },
          "500": {
            "description": "Failed to get the notifs of user_id from the database"
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Delete a notification by it's id",
        "description": "delete notif the database",
        "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer"
        }],
        "responses": {
          "200": {
            "description": "Successfully deleted the notif",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notifications"
              }
            }
          },
          "500": {
            "description": "Failed to delete the notif from the database"
          }
        }
      }
    },
    "/notifications/{id}/receiver": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get all notification of a user",
        "description": "Get notifications where id is receiver",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved notifications",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notifications"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve notifications from the database"
          }
        }
      }
    },
    "/notifications/{id}/update_read": {
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Set a notification as read",
        "description": "Update notification id's read status in the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }],
        "responses": {
          "200": {
            "description": "Successfully updated the update_at of the notification",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notifications"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve notifications from the database"
          }
        }
      }
    },
    "/messages": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "send a message to a user",
        "description": "Create message in the database",
        "parameters": [
          {
            "name": "sender_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "receiver_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "message",
            "in": "formData",
            "required": true,
            "type": "string"
          }],
        "responses": {
          "200": {
            "description": "Successfully created message",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Messages"
              }
            }
          },
          "500": {
            "description": "Failed to create message in the database"
          }
        }
      },
      "delete": {
        "tags": [
          "messages"
        ],
        "summary": "delete a conversation",
        "description": "Delete conversation",
        "parameters": [
          {
            "name": "sender_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "receiver_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          }],
        "responses": {
          "200": {
            "description": "Successfully deleted the conversation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Blocks"
              }
            }
          },
          "500": {
            "description": "Failed to delete the conversation"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Notifications": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "sender_id": {
          "type": "integer"
        },
        "receiver_id": {
          "type": "integer"
        },
        "type": {
          "type": "integer"
        },
        "read": {
          "type": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "Messages": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "sender_id": {
          "type": "integer"
        },
        "receiver_id": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Conversations": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "The unique identifier for the conversation"
        },
        "sender_id": {
          "type": "integer",
          "description": "The id of the user who sent the first message in the conversation"
        },
        "receiver_id": {
          "type": "integer",
          "description": "The id of the user who received the first message in the conversation"
        },
        "message_history": {
          "type": "array",
          "description": "An array of message ids from the messages table, representing the message history for this conversation",
          "items": {
            "type": "integer"
          }
        }
      }
    }
  }
}